# documentation/openapi/openapi/openapi.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: BolaBola API
  termsOfService: https://bolabola.smuaws.com/terms/
  contact:
    name: Contact BolaBola Support
    email: support@bolabola.com
    url: http://bolabola.smuaws.com/contact
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  x-logo:
    url: "https://bolabola.com/logo.png"
    altText: BolaBola logo
  description: >
    This is the **BolaBola** API to demonstrate the functionalities of the BolaBola application.

    # Introduction

    The BolaBola API is designed to provide developers with a comprehensive toolset for integrating with our platform, offering access to our core features and functionalities.

    It also demonstrates the use of the OpenAPI specification to document APIs in a clear and concise manner, facilitating easier integration and understanding of the available API endpoints.

    # OpenAPI Specification

    The OpenAPI Specification offers a wide-ranging and language-agnostic interface to RESTful APIs, enabling both humans and computers to discover and understand the capabilities of a service without direct access to the source code or extensive documentation.

    By adhering to the OpenAPI Specification, we aim to provide a seamless and straightforward experience for developers looking to integrate with the BolaBola platform, ensuring that the available services are easily accessible and understandable.
...
servers:
  - url: https://{tenant}.bolabola.smuaws.com/api/v1
...
paths:
  "/users/{userId}":
    get:
      tags:
        - Users
      summary: Retrieve User Information
      description: Fetches the details of a user given their unique userId. This includes basic user information along with associated tickets.
      parameters:
        - name: userId
          in: path
          required: true
          description: The unique identifier of the user whose information is being retrieved.
          schema:
            type: string
      responses:
        "200":
          description: User information retrieved successfully, including any tickets associated with the user.
        "404":
          description: User not found.
    post:
      tags:
        - Users
      summary: Register a New User
      description: Registers a new user in the system. This endpoint requires a JSON payload containing the new user's information.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas/newuser.yml#/NewUser"
      responses:
        "201":
          description: New user has been successfully created.
        "400":
          description: Invalid input provided.
    put:
      tags:
        - Users
      summary: Update Existing User Information
      description: Updates the information of an existing user identified by their userId. Requires a JSON payload with the updated user information.
      parameters:
        - name: userId
          in: path
          required: true
          description: The unique identifier of the user whose information is to be updated.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas/updateuser.yml#/UpdateUser"
      responses:
        "200":
          description: User information updated successfully.
        "404":
          description: User not found.
        "400":
          description: Invalid input provided.
    delete:
      tags:
        - Users
      summary: Remove a User
      description: Deletes a user from the system using their unique userId. This action is irreversible.
      parameters:
        - name: userId
          in: path
          required: true
          description: The unique identifier of the user to be deleted.
          schema:
            type: string
      responses:
        "204":
          description: User deleted successfully.
        "404":
          description: User not found.  
...
components:
  securitySchemes:
    bolabola_auth:
      description: "OAuth2 authentication for accessing BolaBola APIs."
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://bolabola.com/oauth/authorize
          tokenUrl: https://bolabola.com/oauth/token
          scopes:
            "read:matches": View match information
            "write:tickets": Purchase and manage tickets
    api_key:
      description: "API key authentication for accessing BolaBola APIs."
      type: apiKey
      in: header
      name: X-BolaBola-API-Key